+++ crates/lillinput/src/actions/controller.rs
@@ -1,17 +1,10 @@
 //! Controller for actions.
 
-use std::cell::RefCell;
 use std::collections::HashMap;
 use std::convert::TryFrom;
-use std::rc::Rc;
-use std::str::FromStr;
 
-use crate::actions::commandaction::CommandAction;
 use crate::actions::errors::ActionControllerError;
-use crate::actions::i3action::I3Action;
-use crate::actions::{Action, ActionController, ActionEvent, ActionMap, ActionType};
-use crate::Settings;
-use i3ipc::I3Connection;
+use crate::actions::{Action, ActionController, ActionEvent, ActionMap};
 use itertools::Itertools;
 use log::{debug, info, warn};
 use strum::IntoEnumIterator;
@@ -45,11 +38,12 @@ enum Axis {
 }
 
 impl ActionMap {
-    /// Return a new [`ActionController`].
+    /// Return a new [`ActionMap`].
     ///
     /// # Arguments
     ///
-    /// * `settings` - application settings.
+    /// * `threshold` - Minimum threshold for displacement changes.
+    /// * `actions` - List of action for each action event.
     pub fn new(threshold: f64, actions: HashMap<ActionEvent, Vec<Box<dyn Action>>>) -> Self {
         let action_map = ActionMap { threshold, actions };
 
@@ -156,84 +150,13 @@ impl ActionController for ActionMap {
     }
 }
 
-pub fn extract_action_map(
-    settings: &Settings,
-) -> (
-    HashMap<ActionEvent, Vec<Box<dyn Action>>>,
-    Rc<RefCell<Option<I3Connection>>>,
-) {
-    let mut action_map: HashMap<ActionEvent, Vec<Box<dyn Action>>> = HashMap::new();
-    let connection = Rc::new(RefCell::new(None));
-    let mut connection_exists = false;
-
-    // Create the I3 connection if needed.
-    if settings
-        .actions
-        .values()
-        .flatten()
-        .any(|s| s.type_ == ActionType::I3.to_string())
-    {
-        let new_connection = match I3Connection::connect() {
-            Ok(mut conn) => {
-                info!(
-                    "i3: connection opened (with({:?})",
-                    conn.get_version().unwrap().human_readable
-                );
-                connection_exists = true;
-
-                Some(conn)
-            }
-            Err(error) => {
-                warn!("i3: could not establish a connection: {:?}", error);
-                None
-            }
-        };
-
-        // Update the connection.
-        let connection_option = &mut *connection.borrow_mut();
-        *connection_option = new_connection;
-    }
-
-    // Populate the fields for each `ActionEvent`.
-    for action_event in ActionEvent::iter() {
-        if let Some(arguments) = settings.actions.get(&action_event.to_string()) {
-            let mut actions_list: Vec<Box<dyn Action>> = vec![];
-
-            for value in arguments.iter() {
-                // Create the new actions.
-                match ActionType::from_str(&value.type_) {
-                    Ok(ActionType::Command) => {
-                        actions_list.push(Box::new(CommandAction::new(value.command.clone())))
-                    }
-                    Ok(ActionType::I3) => {
-                        if connection_exists {
-                            actions_list.push(Box::new(I3Action::new(
-                                value.command.clone(),
-                                Rc::clone(&connection),
-                            )));
-                        } else {
-                            warn!("Disabling action as i3 connection could not be established: {value}");
-                        }
-                    }
-                    Err(_) => {
-                        warn!("Unknown action type: '{}", value.type_);
-                    }
-                }
-            }
-
-            action_map.insert(action_event, actions_list);
-        }
-    }
-
-    (action_map, connection)
-}
-
 #[cfg(test)]
 mod test {
-    use crate::actions::controller::{ActionController, ActionEvent, ActionMap, Settings};
+    use crate::actions::controller::{ActionController, ActionEvent, ActionMap};
     use crate::actions::errors::ActionControllerError;
     use crate::extract_action_map;
     use crate::test_utils::default_test_settings;
+    use crate::Settings;
 
     #[test]
     /// Test the handling of an event `finger_count` argument.
